+++ CHANGELOG.md
@@ -5,12 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-[Unreleased]: https://github.com/lerna-stack/akka-entity-replication/compare/v2.2.0...master
+[Unreleased]: https://github.com/lerna-stack/akka-entity-replication/compare/v2.3.0...master
+
+### Changed
+- Improve snapshot synchronization diagnostic logging
+  [PR#216](https://github.com/lerna-stack/akka-entity-replication/pull/216)
+
+### Fixed
+- Consecutive snapshot synchronizations could not synchronize new entity snapshots
+  [#215](https://github.com/lerna-stack/akka-entity-replication/issues/215),
+  [PR#217](https://github.com/lerna-stack/akka-entity-replication/pull/217)
+
+## [v2.3.0] - 2023-06-19
+[v2.3.0]: https://github.com/lerna-stack/akka-entity-replication/compare/v2.2.0...v2.3.0
 
 ### Added
 - Support deleting old events and snapshots
   [#201](https://github.com/lerna-stack/akka-entity-replication/issues/201),
-  [PR#204](https://github.com/lerna-stack/akka-entity-replication/pull/204)
+  [PR#204](https://github.com/lerna-stack/akka-entity-replication/pull/204),
+  [PR#212](https://github.com/lerna-stack/akka-entity-replication/pull/212)
 
 ### Changed
 - Rollback deletes only tagged events after the given sequence number